+++ src/main/java/com/pioneer/medgo/domain/NoticeDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class NoticeDTO {
 	private int id;
-	private int pharmacyid;
+	private Long pharmacyid;
 	private String title;
 	private String content;
 	private Timestamp createdat;
